fix: :bug: Moved Skeleton components to folders and improved background colouring
