+++ src/modules/MangaDetailsPage/components/Skeletons/ChaptersSkeleton.tsx
@@ -1,0 +1,16 @@
+export default function ChaptersSkeleton() {
+  return (
+    <>
+      <button className="self-start bg-gradient-to-r from-gray-700 to-gray-600 w-20 h-6 rounded-md p-2 m-4 gap-1 flex flex-nowrap items-center justify-around"></button>
+      <div className="flex flex-wrap gap-4 justify-center max-w-full m-4 space-x-6">
+        <div className="w-10 h-5 bg-gradient-to-r from-gray-700 to-gray-600 rounded-md"></div>
+        <div className="w-10 h-5 bg-gradient-to-r from-gray-700 to-gray-600 rounded-md"></div>
+        <div className="w-10 h-5 bg-gradient-to-r from-gray-700 to-gray-600 rounded-md"></div>
+        <div className="w-10 h-5 bg-gradient-to-r from-gray-700 to-gray-600 rounded-md"></div>
+        <div className="w-10 h-5 bg-gradient-to-r from-gray-700 to-gray-600 rounded-md"></div>
+        <div className="w-10 h-5 bg-gradient-to-r from-gray-700 to-gray-600 rounded-md"></div>
+        <div className="w-10 h-5 bg-gradient-to-r from-gray-700 to-gray-600 rounded-md"></div>
+      </div>
+    </>
+  );
+}
